requirements.txt
certifi>=2018.11.29
Click>=7.0
gunicorn>=19.9.0
itsdangerous>=1.1.0
Jinja2>=2.10.1
numpy>=1.15.4
opencv-python>=4.0
pillow>=6.2.0
pytesseract>=0.2.6
werkzeug>=0.15.3
astroid>=2.3.1
Flask>=1.1.1
Flask-AutoIndex>=0.6.4
Flask-Silk>=0.2
future>=0.17.1
isort>=4.3.21
lazy-object-proxy>=1.4.2
MarkupSafe>=1.1.1
mccabe>=0.6.1
pylint>=2.4.2
six>=1.12.0
typed-ast>=1.4.0
Werkzeug>=0.16.0
wrapt>=1.11.2
func_timeout>=4.3.5
textblob>=0.17.1
loguru>=0.6.0